requirements.txt
django
django-ninja
django-redis
celery==5.2.3
django-celery-results
django-celery-beat
redis
channels
channels_redis
mysqlclient
requests
unidecode
django-extensions
orjson
uvicorn
websockets
graphviz
pygraphviz
pydot
python-dotenv
pillow
django-cors-headers